factor out deal base
